quic: splitting into client and server
