+++ source/common/http/BUILD
@@ -69,7 +69,7 @@ envoy_cc_library(
         "//source/common/network:filter_lib",
         "//source/common/runtime:runtime_features_lib",
     ] + envoy_select_enable_http3([
-        "//source/common/quic:codec_lib",
+        "//source/common/quic:client_codec_lib",
     ]),
 )
 
@@ -340,6 +340,7 @@ envoy_cc_library(
         "//source/common/http/match_delegate:config",
         "//source/common/network:proxy_protocol_filter_state_lib",
         "//source/common/network:utility_lib",
+        "//source/common/quic:quic_server_factory_stub_lib",
         "//source/common/router:config_lib",
         "//source/common/router:scoped_rds_lib",
         "//source/common/stats:timespan_lib",
