+++ source/common/quic/BUILD
@@ -174,12 +174,10 @@ envoy_cc_library(
     hdrs = ["client_connection_factory_impl.h"],
     tags = ["nofips"],
     deps = [
-        ":active_quic_listener_lib",
         ":envoy_quic_alarm_factory_lib",
         ":envoy_quic_client_session_lib",
         ":envoy_quic_connection_helper_lib",
         ":envoy_quic_proof_verifier_lib",
-        ":envoy_quic_server_session_lib",
         ":envoy_quic_utils_lib",
         "//envoy/http:codec_interface",
         "//envoy/http:persistent_quic_info_interface",
@@ -191,14 +189,34 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "codec_lib",
-    srcs = ["codec_impl.cc"],
-    hdrs = ["codec_impl.h"],
+    name = "client_codec_lib",
+    srcs = ["client_codec_impl.cc"],
+    hdrs = [
+        "client_codec_impl.h",
+        "codec_impl.h",
+    ],
     tags = ["nofips"],
     deps = [
         ":envoy_quic_client_session_lib",
+        ":envoy_quic_utils_lib",
+        "//envoy/http:codec_interface",
+        "//envoy/registry",
+        "@com_github_google_quiche//:quic_core_http_spdy_session_lib",
+    ],
+)
+
+envoy_cc_library(
+    name = "server_codec_lib",
+    srcs = ["server_codec_impl.cc"],
+    hdrs = [
+        "codec_impl.h",
+        "server_codec_impl.h",
+    ],
+    tags = ["nofips"],
+    deps = [
         ":envoy_quic_server_session_lib",
         ":envoy_quic_utils_lib",
+        ":quic_server_factory_stub_lib",
         "//envoy/http:codec_interface",
         "//envoy/registry",
         "@com_github_google_quiche//:quic_core_http_spdy_session_lib",
@@ -251,9 +269,9 @@ envoy_cc_library(
     ],
     tags = ["nofips"],
     deps = [
-        ":envoy_quic_crypto_stream_factory_lib",
         ":envoy_quic_proof_source_lib",
         ":envoy_quic_server_connection_lib",
+        ":envoy_quic_server_crypto_stream_factory_lib",
         ":envoy_quic_stream_lib",
         ":envoy_quic_utils_lib",
         ":quic_filter_manager_connection_lib",
@@ -261,7 +279,7 @@ envoy_cc_library(
         "//source/common/buffer:buffer_lib",
         "//source/common/common:assert_lib",
         "//source/common/http:header_map_lib",
-        "@com_github_google_quiche//:quic_core_http_spdy_session_lib",
+        "@com_github_google_quiche//:quic_server_http_spdy_session_lib",
     ],
 )
 
@@ -278,7 +296,7 @@ envoy_cc_library(
     tags = ["nofips"],
     deps = [
         ":envoy_quic_client_connection_lib",
-        ":envoy_quic_crypto_stream_factory_lib",
+        ":envoy_quic_client_crypto_stream_factory_lib",
         ":envoy_quic_proof_verifier_lib",
         ":envoy_quic_stream_lib",
         ":envoy_quic_utils_lib",
@@ -350,9 +368,9 @@ envoy_cc_library(
     hdrs = ["envoy_quic_dispatcher.h"],
     tags = ["nofips"],
     deps = [
-        ":envoy_quic_crypto_stream_factory_lib",
         ":envoy_quic_proof_source_lib",
         ":envoy_quic_server_connection_lib",
+        ":envoy_quic_server_crypto_stream_factory_lib",
         ":envoy_quic_server_session_lib",
         ":quic_stat_names_lib",
         "//envoy/network:listener_interface",
@@ -437,10 +455,10 @@ envoy_cc_library(
     alwayslink = LEGACY_ALWAYSLINK,
 )
 
-# Create a single target that contains all the libraries that register factories.
+# Create a single target that contains all the libraries that register client factories.
 # All of these are needed for this extension to function.
 envoy_cc_library(
-    name = "quic_factory_lib",
+    name = "quic_client_factory_lib",
     tags = ["nofips"],
     # QUICHE can't build against FIPS BoringSSL until the FIPS build
     # is on a new enough version to have QUIC support. Remove it from
@@ -449,7 +467,38 @@ envoy_cc_library(
         "//bazel:boringssl_fips": [],
         "//bazel:boringssl_disabled": [],
         "//conditions:default": [
-            ":codec_lib",
+            ":client_codec_lib",
+            ":quic_transport_socket_factory_lib",
+            "//source/extensions/quic/proof_source:envoy_quic_proof_source_factory_impl_lib",
+        ],
+    }),
+)
+
+# The factory for server connection creation.
+envoy_cc_library(
+    name = "quic_server_factory_stub_lib",
+    hdrs = ["server_connection_factory.h"],
+    deps = [
+        "//envoy/http:codec_interface",
+        "//envoy/network:connection_interface",
+        "//source/common/http/http3:codec_stats_lib",
+        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
+    ],
+)
+
+# Create a single target that contains all the libraries that register factories.
+# All of these are needed for this extension to function.
+envoy_cc_library(
+    name = "quic_server_factory_lib",
+    tags = ["nofips"],
+    # QUICHE can't build against FIPS BoringSSL until the FIPS build
+    # is on a new enough version to have QUIC support. Remove it from
+    # the build until then. Re-enable as part of #7433.
+    deps = select({
+        "//bazel:boringssl_fips": [],
+        "//bazel:boringssl_disabled": [],
+        "//conditions:default": [
+            ":server_codec_lib",
             ":quic_transport_socket_factory_lib",
             "//source/extensions/quic/crypto_stream:envoy_quic_crypto_server_stream_lib",
             "//source/extensions/quic/proof_source:envoy_quic_proof_source_factory_impl_lib",
@@ -499,12 +548,24 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "envoy_quic_crypto_stream_factory_lib",
-    hdrs = ["envoy_quic_crypto_stream_factory.h"],
+    name = "envoy_quic_client_crypto_stream_factory_lib",
+    hdrs = ["envoy_quic_client_crypto_stream_factory.h"],
+    tags = ["nofips"],
+    deps = [
+        "//envoy/config:typed_config_interface",
+        "@com_github_google_quiche//:quic_client_session_lib",
+        "@com_github_google_quiche//:quic_core_http_spdy_session_lib",
+    ],
+)
+
+envoy_cc_library(
+    name = "envoy_quic_server_crypto_stream_factory_lib",
+    hdrs = ["envoy_quic_server_crypto_stream_factory.h"],
     tags = ["nofips"],
     deps = [
         "//envoy/config:typed_config_interface",
         "@com_github_google_quiche//:quic_core_http_spdy_session_lib",
+        "@com_github_google_quiche//:quic_server_session_lib",
     ],
 )
