+++ source/common/quic/envoy_quic_client_session.h
@@ -3,8 +3,8 @@
 #include "envoy/http/http_server_properties_cache.h"
 
 #include "source/common/quic/envoy_quic_client_connection.h"
+#include "source/common/quic/envoy_quic_client_crypto_stream_factory.h"
 #include "source/common/quic/envoy_quic_client_stream.h"
-#include "source/common/quic/envoy_quic_crypto_stream_factory.h"
 #include "source/common/quic/quic_filter_manager_connection_impl.h"
 #include "source/common/quic/quic_stat_names.h"
