+++ source/common/quic/server_codec_impl.cc
@@ -1,6 +1,5 @@
-#include "source/common/quic/codec_impl.h"
+#include "source/common/quic/server_codec_impl.h"
 
-#include "source/common/quic/envoy_quic_client_stream.h"
 #include "source/common/quic/envoy_quic_server_stream.h"
 
 namespace Envoy {
@@ -12,11 +11,6 @@ namespace Quic {
 EnvoyQuicStream* quicStreamToEnvoyStream(quic::QuicStream* stream) {
   return dynamic_cast<EnvoyQuicStream*>(stream);
 }
-EnvoyQuicClientStream* quicStreamToEnvoyClientStream(quic::QuicStream* stream) {
-  return dynamic_cast<EnvoyQuicClientStream*>(stream);
-}
-
-bool QuicHttpConnectionImplBase::wantsToWrite() { return quic_session_.bytesToSend() > 0; }
 
 QuicHttpServerConnectionImpl::QuicHttpServerConnectionImpl(
     EnvoyQuicServerSession& quic_session, Http::ServerConnectionCallbacks& callbacks,
@@ -58,50 +52,7 @@ void QuicHttpServerConnectionImpl::goAway() {
   quic_server_session_.SendHttp3GoAway(quic::QUIC_PEER_GOING_AWAY, "server shutdown imminent");
 }
 
-QuicHttpClientConnectionImpl::QuicHttpClientConnectionImpl(
-    EnvoyQuicClientSession& session, Http::ConnectionCallbacks& callbacks,
-    Http::Http3::CodecStats& stats,
-    const envoy::config::core::v3::Http3ProtocolOptions& http3_options,
-    const uint32_t max_request_headers_kb, const uint32_t max_response_headers_count)
-    : QuicHttpConnectionImplBase(session, stats), quic_client_session_(session) {
-  session.setCodecStats(stats);
-  session.setHttp3Options(http3_options);
-  session.setHttpConnectionCallbacks(callbacks);
-  session.setMaxIncomingHeadersCount(max_response_headers_count);
-  session.set_max_inbound_header_list_size(max_request_headers_kb * 1024);
-}
-
-void QuicHttpClientConnectionImpl::goAway() {
-  quic_client_session_.SendHttp3GoAway(quic::QUIC_PEER_GOING_AWAY, "client goaway");
-}
-
-Http::RequestEncoder&
-QuicHttpClientConnectionImpl::newStream(Http::ResponseDecoder& response_decoder) {
-  EnvoyQuicClientStream* stream =
-      quicStreamToEnvoyClientStream(quic_client_session_.CreateOutgoingBidirectionalStream());
-  ASSERT(stream != nullptr, "Fail to create QUIC stream.");
-  stream->setResponseDecoder(response_decoder);
-  if (quic_client_session_.aboveHighWatermark()) {
-    stream->runHighWatermarkCallbacks();
-  }
-  return *stream;
-}
-
-void QuicHttpClientConnectionImpl::onUnderlyingConnectionAboveWriteBufferHighWatermark() {
-  quic_client_session_.PerformActionOnActiveStreams([](quic::QuicStream* quic_stream) {
-    ENVOY_LOG(debug, "runHighWatermarkCallbacks on stream {}", quic_stream->id());
-    quicStreamToEnvoyStream(quic_stream)->runHighWatermarkCallbacks();
-    return true;
-  });
-}
-
-void QuicHttpClientConnectionImpl::onUnderlyingConnectionBelowWriteBufferLowWatermark() {
-  quic_client_session_.PerformActionOnActiveStreams([](quic::QuicStream* quic_stream) {
-    ENVOY_LOG(debug, "runLowWatermarkCallbacks on stream {}", quic_stream->id());
-    quicStreamToEnvoyStream(quic_stream)->runLowWatermarkCallbacks();
-    return true;
-  });
-}
+REGISTER_FACTORY(QuicHttpServerConnectionFactoryImpl, QuicHttpServerConnectionFactory);
 
 } // namespace Quic
 } // namespace Envoy
