+++ source/extensions/quic/crypto_stream/BUILD
@@ -26,7 +26,7 @@ envoy_cc_library(
     ],
     deps = [
         "//envoy/registry",
-        "//source/common/quic:envoy_quic_crypto_stream_factory_lib",
+        "//source/common/quic:envoy_quic_server_crypto_stream_factory_lib",
         "@envoy_api//envoy/extensions/quic/crypto_stream/v3:pkg_cc_proto",
     ],
     alwayslink = LEGACY_ALWAYSLINK,
@@ -60,7 +60,7 @@ envoy_cc_library(
         "//test:__subpackages__",
     ],
     deps = [
-        "//source/common/quic:envoy_quic_crypto_stream_factory_lib",
+        "//source/common/quic:envoy_quic_client_crypto_stream_factory_lib",
     ],
     alwayslink = LEGACY_ALWAYSLINK,
 )
